Creates session folder before doing something
modified:   sessions.py

+++ sessions.py
@@ -7,11 +7,13 @@ def get_session(session_name):
   return session
 
 def get_session_list():
-  sessions = {}
-  for root, dirnames, files in os.walk("sessions"):
-    for session_name in files:
-      sessions[session_name] = load_session(session_name)
-  return sessions
+  if exists_or_create_session_folder():
+    sessions = {}
+    for root, dirnames, files in os.walk("sessions"):
+      for session_name in files:
+        sessions[session_name] = load_session(session_name)
+    return sessions
+  return {}
 
 def create_session(session_name, data):
   session = ConfigParser.ConfigParser()
@@ -37,20 +39,32 @@ def add_order_to_session(session_name, order):
   save_session(session_name, session)
 
 def load_session(session_name):
-  session = ConfigParser.ConfigParser()
-  file_path = os.path.join("sessions", session_name)
-  with open(file_path, "r") as fp:
-    session.readfp(fp)
-  return session
+  if exists_or_create_session_folder():
+    session = ConfigParser.ConfigParser()
+    file_path = os.path.join("sessions", session_name)
+    with open(file_path, "r") as fp:
+      session.readfp(fp)
+    return session
 
 def save_session(session_name, session):
-  file_path = os.path.join("sessions", session_name)
-  if os.path.isfile(file_path):
-    return False
-  else:
-    with open(file_path, "w") as fp:
-      session.write(fp)
-    return True
+  if exists_or_create_session_folder():
+    file_path = os.path.join("sessions", session_name)
+    if os.path.isfile(file_path):
+      return False
+    else:
+      with open(file_path, "w") as fp:
+        session.write(fp)
+      return True
+  return False
+
+def exists_or_create_session_folder():
+  if not os.path.isdir("sessions"):
+    try:
+      os.mkdir("sessions")
+    except:
+      print("failed to create sessions folder")
+      return False
+  return True
 
 def sort_sessions_by_deadline(sessions):
   # TODO: add sorting by time
